Add oneof as a word
This adds `oneof` as a valid word so `Oneof` isn't flagged as a typo.
I wasn't sure what the correct capitalization should be since the RFC
styles it as "Oneof Input Objects" but in the schema it is
`__Type.oneOf`. I tried to use "Oneof" in prose and `oneOf` in code.